Encode signing scheme when generating signature
